[Clang] Add support for STDC CX_LIMITED_RANGE pragma.
This pragma is defined in the C specification. Notably, the C specification
makes the "DEFAULT" value identical to "OFF" for this pragma, unlike other
floating-point pragmas which are undefined behavior. This may be surprising to
some users.

This patch builds on, and requires, the complex intrinsics to make the pragmas
work properly.

Depends on D119290

Differential Revision: https://reviews.llvm.org/D119291